Combined tests in test_preprocess and test_preprocessor into test_preprocess.py
